Split the auth-shape classifiers out of transport into runtime/
hasStaticCredentials and classifyAuthShape are pure, config-derived
classification with no request or transport state, but they lived in
transport/bearerGuard.ts — so runtime/banner.ts and runtime/wikiCapability.ts
imported up into transport to reach them.

Move them (with the AuthShape/Transport types) to a new runtime/authShape.ts, a
lower-layer primitive. transport/bearerGuard.ts keeps only the boot-time guard
(evaluateBearerGuard) and now imports hasStaticCredentials down from runtime.
Runtime no longer reaches into transport for auth-shape classification. Split
the test file to mirror the source tree; every original case is preserved. Pure
move, no behaviour change.
