+++ src/runtime/wikiCapability.ts
@@ -3,7 +3,7 @@ import type { ToolContext } from './context.js';
 import type { ExtensionPack } from '../tools/extensions/types.js';
 import { extensionPacks } from '../tools/extensions/index.js';
 import { getRuntimeToken } from './requestContext.js';
-import { hasStaticCredentials } from '../transport/bearerGuard.js';
+import { hasStaticCredentials } from './authShape.js';
 
 const CORE_WRITE_TOOL_NAMES: readonly string[] = [
 	'create-page',
